requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
pkgutil-resolve-name==1.3.10
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
rpds-py==0.17.1
SQLAlchemy==2.0.25
typing-extensions==4.9.0
werkzeug==3.0.1
zipp==3.17.0
